requirements.txt
anthropic>=0.21.0
prompt-toolkit>=3.0.0
python-dotenv>=1.0.0
